+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.0.6
-- Idk yet
+- Remove color flashing for menus, only have it for in-game objects.  
+(Note that this is the first version released on the Geode index.)
 
 # 1.0.5
 - Make it so the blinking eye slowly moves into the screen before blinking, so you know it's coming.